+++ prep/src/cmd/clippy.rs
@@ -5,14 +5,25 @@ use anyhow::{Context, ensure};
 
 use crate::cmd::CargoTargets;
 use crate::session::Session;
-use crate::tools::cargo;
+use crate::tools::cargo::{Cargo, CargoDeps};
 use crate::ui;
 
 /// Runs Clippy analysis on the given `targets`.
 ///
-/// In `strict` mode warnings are treated as errors.
-pub fn run(session: &Session, targets: CargoTargets, strict: bool) -> anyhow::Result<()> {
-    let mut cmd = cargo::new("")?;
+/// In `strict` mode warnings are treated as errors and Cargo version is locked.
+pub fn run(session: &mut Session, strict: bool, targets: CargoTargets) -> anyhow::Result<()> {
+    let mut cmd = if strict {
+        let tools_cfg = session.config().tools();
+        let rustup_ver_req = tools_cfg.rustup().clone();
+        let ver_req = tools_cfg.rust().clone();
+        let toolset = session.toolset();
+        let deps = CargoDeps::new(rustup_ver_req);
+        toolset.get::<Cargo>(&deps, &ver_req)?
+    } else {
+        let toolset = session.toolset();
+        let deps = CargoDeps::new(None);
+        toolset.get::<Cargo>(&deps, None)?
+    };
     let mut cmd = cmd
         .current_dir(session.root_dir())
         .arg("clippy")